+++ NEWS.md
@@ -1,4 +1,7 @@
 
+# hutils 0.10.0
+* New function `implies`, logical implies.
+
 # hutils 0.9.0
 * Changed `if_else` to reflect dplyr's formals so it can be a drop-in replacement.
 * Fix error for multi-length `missing` value in `if_else` when length-one `condition`.